requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
flask-sock==0.6.0
Flask-SocketIO==5.3.4
gunicorn==20.1.0
h11==0.14.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
simple-websocket==0.10.1
Werkzeug==2.3.4
wsproto==1.2.0
zipp==3.15.0


flask-socketio
Pillow==9.4.0